Makefile: Allow PLATFORM_VERSION variable to be changed in make call
In commit 8215689277d ("Makefile: Set `PLATFORM_VERSION`"), a new
variable called `PLATFORM_VERSION` is defined directly in Makefile.
However, the variable was set with "simply expanded" sign (":="), which
will set the value even if it has been set from another `make` call.
Instead, we can use "?=" for the assignment sign. With this, Linux build
system will not override the said variable.

Change the assignment sign to "?=" to allow `PLATFORM_VERSION` variable
to be changed.

Signed-off-by: Hendra Manudinata <hendra@manoe.my.id>